Style error messages for select fields
move adding of error class span to error message to errors_text method
so everything can use it.

Add errors class to wrapping p for select method. Should probably do it
to the rest too, but wary of touching too much in this gem as not sure
what else uses it.